third_party/blink/web_tests/platform/generic/fast/dom/shadow/get-element-by-id-in-shadow-mutation-expected.txt

Tests to ensure that ShadowRoot.getElementById works even after mutation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS null is shadow.getElementById('foo')
PASS e0a is shadow.getElementById('foo')
PASS e0b is shadow.getElementById('foo')
PASS e0c is shadow.getElementById('foo')
PASS e0b is shadow.getElementById('foo')
PASS e0a is shadow.getElementById('foo')
PASS null is shadow.getElementById('foo')
PASS successfullyParsed is true

TEST COMPLETE
